Move Result extension to an existing file
