--- shared/src/commonMain/kotlin/cz/matee/devstack/kmp/shared/base/Result+Extensions.kt
@@ -1,4 +1,0 @@
-package cz.matee.devstack.kmp.shared.base
-
-fun <T : Any> Result<T>.getOrNull(): T? =
-    (this as? Result.Success)?.data
